feat: viewDidLoad input 생성 및 메서드 이동
event, news 및 트래킹 요청 이동 / 홈 정보 가져오는 기능 구현
